Enhance notification badge logic to allow adding more of them
Add NotificationBadge type to include all types of notification badge
(red dots) that could exist in the app.

Add NotificationBadgeService, with an API that allows to update the
value in a suspending manner or listen to it as a flow. Flow.first()
can be used to simply ask for the current value inside a coroutine.

Migrate seen cross sells from being saved into a string set inside
Datastore preferences to individual items with the key existing inside
the connected NotificationBadge. Saved as a simple boolean, true for
seen, false for not seen yet.

Add showNotificationBadge field in InsuranceModel.CrossSellHeader

Update TabNotificationService to internally make use of
NotificationBadgeService instead of the dataStore directly.
TabNotificationService is kept to provide a simpler API for the
LoggedInViewModelImpl to consume