Add separate CrossSellHeader viewHolder item with an animated badge
Add custom DiffUtil to properly show the change of the notification
badge going from unseen to seen. Setting supportsChangeAnimations as
false will actually let only NotificationSubheadingViewHolder animate
the content change itself. This doesn't affect the rest of the items as
areItemsTheSame returns true for them, meaning that a different
animation plays anyway, not the "changeAnimation" that was set to false.