Change from custom FadingTopAppBar into the one provided by accompanist
Accompanist-ui provides a TopAppBar that also exposes a contentPadding
parameter, making it possible to use a TopAppBar while also giving it
custom inset handling. This way it's possible to make it fading simply
by providing it a different alpha as its background color while also
keeping the proper specs for the material top app bar.