Remove termination info which was showing faulty information
Context:
https://hedviginsurance.slack.com/archives/CPCUHMMJQ/p1635932572116000
Optimally this should be shown in the case where specifically a previous
contract was over after the member went through the moving flow, but
there is no way to know that when we get a new clean contract, therefore
it was decided to show nothing instead.
The information about the old contract being active 30 days after the
moving is presented at the end of the moving flow.