Add explicit dependency to compose runtime and foundation
Bump foundation to an alpha version to make SelectActionView actually
show. There seems to be a bug that does not show a composable when
inside a androidx.compose.ui.platform.ComposeView or potentially due to
some other combination of configurations not entirely clear.

Explicitness is needed to avoid problems like what I had while fixing
this bug. Bumping mdcAdapter, a seemingly unrelated dependency was
fixing the bug due to it pulling the beta version of compose foundation.

Adding an explicit version of runtime will also serve as a great way to
avoid potential future problems with libraries bringing it in
transitively and fixing or breaking things in silence.

Before merging this, probably a good idea to wait for compose 1.1.0 to
become stable