Make AudioPlayer both initializable and closeable on demand
Since MediaPlayer preforms a network request for the audio file on
initialization, we should be wary of creating it more than needed.
Instead move the initialization part to a function which can be called
later (in a DisposableEffect in this case) when we know for sure that
we've completed a successful composition.