Alter safeQuery to handle error responses
When graphQL returns an ok 200 result with partial data but an error in
the errors body, before it was considered a successful response,
breaking some embark flows. This change makes it so that such responses
are interpreted as a failed response, making the logic continue on the
unhappy path instead.