Make offer query fetch the quoteIds from the latest selection
The problem was as follows. The getList/get functions first look into
the stored values and if not existing then look into for example the
`stage` list. This means that if the key exists in both places, the
stored values are looked into first and the ones in staging are
disregarded.

This solves the bug where one may go through the entire flow, reach the
offer page. Then go back by pressing the back button to go and change
some input, and then continue the flow again. Then in the offer stage,
the old saved value is used instead of the latest one.

It was also considered to do a proper commit at this stage and a pop
after we're back into it, but was considered not a possible solution due
 to how some passages are skipped when coming back like OfferSuccess