Make address selection more fluid
Before on every address selection, we did the same network request for
the address autocomplete twice. Once for populating the list and once
for getting the final address. These then individually reported to the
viewState through a `combine` of their flows, meaning that the results
came in an unexpected order. Specifically, almost all the time, they
list result would come either first or so close to the final address
result that the list was updating and then instantly after that the
screen would dismiss the screen. This made the UI look jumpy when
clicking on the final address as explained here
https://hedvig.atlassian.net/browse/APP-1154?focusedCommentId=11349

This change makes it so that this process is all streamlined from the
same source, `currentInput`. `withHistoryOfLastValue` is utilized to
get access to the previous value to do the necessary logic. Now the two
network requests happen in parallel again, but the final address result
is taking precedence over the result of the normal list updating. This
can work since if we get a final address result we can drop everything
and simply finish the activity with that as the result. To make the UI
not jump in that case, we simply retain the old list items and we
populate the finalAddress field which the UI then uses to finish the
activity.