Make MainCoroutineRule set StandardTestDispatcher as the main dispatcher
To my understanding, because ViewModel uses Dispatchers.Main.immediate
by default we need to set the main dispatcher to our own TestDispatcher
to make it function as expeced.

Aside from StandardTestDispatcher, UnconfinedTestDispatcher also exists
which comes with some implicit behavior which may not be what we want.
https://kotlinlang.slack.com/archives/C1CFAFJSK/p1642775389275900
?thread_ts=1642773185.273800&cid=C1CFAFJSK talks about this, which
makes me think we probably want to use the Standard dispatcher and in
cases when needed call `advanceTimeBy`, `advanceUntilIdle` or
`runCurrent` to be explicit about what is happening.