Make use of advanceUntilIdle to await a delayed coroutine
After calling `viewModel.onRetrievePriceInfo()` we call `runCurrent()`
to get inside the first `launch` that happens from viewModelScope since
we need to be explicit of it starting. This in turn turns the isLoading
state to true and calls `startDataCollection`. This however was mocked
to contain a delay of 100ms, so in order to wait for its completion we
can either hardcode an `advanceTimeBy(100)` which is bound to be wrong
in case anything changes, or simply use `advanceTimeUntilIdle` which
recognizes that a coroutine is currently running and awaits until it's
done to continue execution. This means that after that point we're
after the result was returned therefore the viewState must have been
updated and the loading state was set as false.