Fix shouldValidatePersonalIdentificationNumbersForDenmark test
TextHint simply doesn't seem to work, it fails with: "Expected helper
text is , but actual is null" but
R.string.simple_sign_login_text_field_helper_text_dk isn't null.
The only reason this may be happening is that it somehow fetches the
string from the wrong locale since that string has no translation for
locales ["fr", "nb-rNO", "sv-rSE"] and it tries to fetch from one of
those and of course fails to do so.