Some juggling with HAnalytics implementation
Extracts all the network-related functionality to a retrofit-like
interface called HAnalyticsService which can take care of all the
networking logic like constructing the OkHttp request, adding the
headers and so on. Leaves a simple interface of the functions that we
can access from the REST API provided to us from the hAnalytics backend
service.

This allows our implementation of the HAnalytics interface be a simple
delegation to one class which manages what happens when we want to send
an HAnalyticsEvent and one that handles experiments.

ExperimentManager does exactly that, make use of the REST API interface,
keeps a local map of the loaded experiments, and handles all the
necessary concurrency protection that is needed using a Mutex.

SendHAnalyticsEventUseCase exists as the de-facto way to send an
HAnalytics event as it makes sure that all the registered sinks will
receive each event that is sent.