APP-1029: Fix embark compose action buttons not showing before getting clicked (#799)
* Revert "Revert "Merge pull request #768 from HedvigInsurance/fix/APP_467_center_align_embark_action_button" (#798)"

This reverts commit fb2945bc

* Make actionsComposeView gone when not used

* Add explicit dependency to compose runtime and foundation

Bump foundation to an alpha version to make SelectActionView actually
show. There seems to be a bug that does not show a composable when
inside a androidx.compose.ui.platform.ComposeView or potentially due to
some other combination of configurations not entirely clear.

Explicitness is needed to avoid problems like what I had while fixing
this bug. Bumping mdcAdapter, a seemingly unrelated dependency was
fixing the bug due to it pulling the beta version of compose foundation.

Adding an explicit version of runtime will also serve as a great way to
avoid potential future problems with libraries bringing it in
transitively and fixing or breaking things in silence.

Before merging this, probably a good idea to wait for compose 1.1.0 to
become stable

* Improve function name after no longer dealing with RecyclerViews

* Fix details after rebase

* Make material transition not crash with compose view

* Run ktlint

* Remove some leftover optIns that are no longer experimental

* Make action selection grid items support the top badge

* Make grid adjust all items to have the same height

* Fix bugs with wrong sizing

* Add engineering mode badges to check UI

* Fix inner text having double padding horizontally

* Remove an extra bottom padding

* Remove a non-true documentation comment

The minimum height functionality is an implementation detail of one
call site. This isn't true for all possible callers of that layout

* Remove an unecessary part of placing a placeable

This was being taken care of by the calculation that initializes
`layoutHeight` by using the appropriate height for it. No need to do
this extra check anymore.

* Add some removed support for an engineering mode screen