Fix app crashing when going back to LoggedInActivity
When using LiveData, we depended on some subtle behavior of it not
calling the `observe` callback when going back into the activity
somehow. With the change to observing a flow, we now get an update when
we go back to this activity. This would work fine, however bottom nav
bar treats `inflateMenu` as adding the new items *on top* of everything
that it had before without clearing the old menu. Simply clearing it
before doesn't work either, because an animation plays with the new
items being populated despite the fact that they're the same items, on
top of losing track of what was selected before-hands.
In the end, a simple check was added to compare the menu to be inflated
with the previous one which is saved in a field in the activity. If
they're the same we just don't make any changes in the menu.