Make "signQuotesAndClearCache" clear the apollo cache
I am guessing this was something that was deleted by accident and it
was never added back in. This should make it so that this
https://github.com/HedvigInsurance/android/pull/987 is fixed and the
insurances tab doesn't show stale data.

As a side effect, we can now make the GetContractsUseCase be able to
use the cache again and not force it to hit the Network