Provide a Unit test that ensures that constructing a Data object works
Add two side-to-side tests of how it would work with using the apollo3
test builders, and how it would work by constructing the Operation.Data
model by hand.
If the correct __typename is provided for the necessary parts of the
model both ways should work the same.

Particularly, it seems like in the places where we're using inline
fragments we need to be explicit about which __typename we want to
provide. This is not something that can be deduced manually. This
worked in 2.x since the wrong __typename was provided as a default in
the constructor and the parser could still manage to convert the json
into the appropriate object, albeit with the wrong __typename.

Some context:
https://kotlinlang.slack.com/archives/C01A6KM1SBZ/p1653488677389899?thread_ts=1653488305.816139&cid=C01A6KM1SBZ