Send `receivedQuotes` analytics event (#1093)
* Send `receivedQuotes` analytics event

* Add query to fetch only the quote cart Checkout status

* Let SwedishBankIdSignDialog use QuoteCartCheckoutStatusQuery

Also make it work with a view state instead of events

* Remove animateContentSize from compose dialog

When the text becomes bigger it has a jarring effect of jumping back and
forth to being one and two lines with the last emoji character jumping
between being on line 1 and 2. Removing this animations makes it feel
much smoother.

* Create a fast track when the checkout is pending

When trying to get quote Ids and start a sign, check only if the
checkout is already pending and if yes, fast-forward to opening the
bankID dialog without querying the entire quote cart query.