Make apollo use the test dispatcher
This allows apollo network calls to use the test dispatcher to run the
calls. This is needed because when running apollo requests not from the
test itself but inside a viewModel which uses Dispatchers.Main.immediate
as a dispatcher for viewModelScope. Since the two dispatchers were
different, the apollo tasks were never started, even after doing
runCurrent() or advanceUntilIdle() since apollo was using its own
dispatcher.

The `testScheduler` needs to be passed inside the
StandardTestDispatcher() since sometimes we're overriding main with
MainCoroutineRule and it works even without it, but when we don't do
that (when we're not doing ViewModel tests we don't need that rule) then
 it's using a newly created TestCoroutineScheduler which does not allow
 us to fast-forward time through delays and simply does not execute
 suspending functions when run inside the test directly. It simply hangs
 forever instead.

 Some more context:
 https://kotlinlang.slack.com/archives/C01A6KM1SBZ/p1655907577391439?thread_ts=1655813143.939929&cid=C01A6KM1SBZ