Fix charity screen showing the wrong screen
When cashback information was coming as null from the backend as null,
we were mapping it to a CashbackUiState object with all null parameters.
The UI however was checking for if this entire object was null to
instead show the version of the screen where one is allowed to pick
their charity option. This fixes actually mapping to a null
CashbackUiState again so that the UI can show the proper information.