Ignore flaky test until the race condition is addressed
For the test to run reliably we'd need to somehow make sure that the
test waits for the coroutine to finish before asserting about if the
intent was intended yet or not. The problem arises from the fact that
the sunsetting interceptor does let the response go through, and *then*
tries to look inside the response for the necessary information about
whether the sunsetting activity should show up or not. The current test
asserts too fast and does not give it the time to run that piece of code