Change SimpleSign VM to handle the error case too
After discussion, it was deemed better to actually error out on the
error auth case despite the fact that it happens even without any input
on a *very* specific sequence of events. In such cases the retry button
should simply let the user go through the flow again and succeed on the
second try.