Do not try to launch WebOnboarding if not supported
According to the docs
https://developer.android.com/guide/components/intents-common#Browser
And the recent crash (probably from the Play Store monkey testers)
https://console.firebase.google.com/u/0/project/hedvig-app/crashlytics/app/android:com.hedvig.app/issues/3386af3e8dbbd2a0b6e2609c949ffff3?time=last-thirty-days&versions=7.0.0%20(1658995833);7.0.0%20(1658995590)&sessionEventKey=62E24BFC01E8000125FF85939E5C49DF_1703787032309783907

One should not `.startActivity` on an Intent which hasn't been checked
if it works or not before. If the phone has no supporting app that can
launch this intent, in this case a URL, it will straight up crash the
app.

Current change prints NETWORK_ERROR_ALERT_MESSAGE aka "We couldn't
reach Hedvig right now, are you sure you have an internet connection?"
since we don't have a more appropriate string ready, but at least it
might prompt people to go to the website directly