Introduce hanalytics module which encapsulates its internals
Exposes specifically only the things that need to be available in any
dependant modules.

Things like HAnalyticsSink, to introduce more of them if there is such
a need.
The HAnalytics interface from the remote dependency as an `api`
dependency to allow normal analytics events calls.
And FeatureManager for all things experiments