Introduce hanalytics-engineering (+ noop) module
The engineering mode sinks, activity and notification moves to a
separate module to encapsulate all of its functionality.

The -noop module is required so that even when the app is built in
release mode where the symbols from hanalytics-engineering do not
normally exist, we need to be able to reference them. This then exposes
the same public API as the original module, only it is a noop.

Noop module explanation:
https://wajahatkarim.com/2018/05/creating-separate-modules-for-debug-and-release/