Make FakeFeatureManager not make assumptions for return types
With the following approach the developer will now need to be explicit
that if one of the functions from FeatureManager is called what it will
return, failing if it is not set. When the return value needs to be
there but is not important for the present test one can pick a random
return type.
Also make constructor receive lambdas that return a result to improve
its flexibility regarding what it can return instead of one fixed value.