Add `apiValidation` configuration for project.
For going for a "opt-in" approach instead of an "opt-out". Specifically
the configuration ignores all modules except for the ones specified in
the `projectsToRunApiValidationOn` list.

That is so that we can test how this could look like for a module like
`hanalytics` which is supposed to only expose a specific API for the
clients while hiding all implementation details.