Use safeDrawing instead of safeContent
SafeContent also includes the navigation insets which on devices that
use gesture navigation means that there's a quite big gap at the edges
so as to avoid putting content on top of the places where the user may
want to do their navigation gestures.
In our case we do not mind this as we do not have clickable content at
the edges anyway.