Add another ignore for ktlint for function-signature
This is for functions which have their parameters in separate line but
technically could also be in one line since the available space allows
it. I am disabling it since sometimes for clarity we do want to have
that kind of formatting so that the function is easier to read.

With that said, the IDE currently does not handle multiple
disabled_rules which contain a colon ':' in their name, so the IDE will
complain, but it's a false positive. The ktlint plugin can handle this
just fine.
Relevant issue here: https://issuetracker.google.com/issues/242824740