Make InsuranceAdapter viewholders not impl smth they don't need
As a step to pass an imageLoader to a child I figured I wouldn't want to
pass this in the bind function since not all viewholders need this
imageLoader. Same should apply to the marketManager as well. Or anything
else that is VH specific, like the retry function. Those can be passed
in the VH constructor instead.