Properly catch and map watch errors
The previous try catch was on the flow creation aka it was never
catching the exceptions thrown inside the flow, and it made it through
to the call site. This change makes it so that the errors are caught and
forwarded to the call site after being mapped into a QueryResult as it
should be doing in the first place.