Remove error message from error item
The user-facing error with this ends up being stuff like "Multiple
exceptions happened", something that we devs may find useful but not so
much for the person using the app. This commit removes propagating this
message further than the ViewModel and lets the UI pick the error
message itself. In this case, usually the home_tab_error_body aka "check
your internet connection".