Change referrals ViewModel to handle no internet case
Main changes is that it now tries to watch for the query, and if it
fails to fetch it the *first* time, instead of waiting on the `watch`
forever and never completing, it now throws inside the watch flow and
gets converted to a proper error state. Then to allow retrying again,
the error state contains the retry button along with the swipe to
refresh functionality.