Remove ignored dependencies from renovate
The reason we were ignoring updates for kotlin is that they were tied to
compose compiler version which made it so that we had to delay by a lot
when we eventually did these bumps. With Compose compiler 1.3 and
onwards this is no longer an issue therefore this can safely be removed
now and we can keep up to date more easily.