Create :notification-badge-data
Moves all notification-badge related logic into this module, only
exposing the two things that the rest of the app is interested about.
The `TabNotificationBadgeService` and the
`CrossSellCardNotificationBadgeService`. One to interface with the
badges in the bottom nav bar, and one for the "more insurances" section
of the insurance tab.

This was done to simplify the interface to the rest of the app, but
started as an effort to fix a bug where one of the badges was not being
cleared, and the added tests check for this scenario.