Turbine 0.10 uses unconfined dispatcher so no need to runCurrent
Also means that the in-between states are not lost which means in the
previous version of the tests there was a duplicate emission which the
tests did not account for.

In this case the datastore changing meant that both
`crossSellBottomNavNotificationBadgeService.showNotification()` and
`referralsNotificationBadgeService.showNotification()` were getting an
updated value and `combine` was triggering these changes one *after*
the other resulting in the turbine flow receiving the empty result
twice. Since the referrals information was the same it is surprising
that it emitted again in the first place, but it seems like datastore
does not do anything to avoid emitting the same data twice, so a simple
`.distinctUntilChanged` does the job in our case.