Remove Showkase
We're really not using this atm plus it's introducing kapt into our
build. If we are to use it again we should go with the KSP version. But
since we're not using it, it's easier to simply delete it for the time
being, we can always revert this commit when we reintroduce it.