Make settings screen not show "not set" initially
So the way that we use this preferences screen is kind of odd, since we
do not actually store a bunch of this information inside shared
preferences, so these changes may look a bit odd.
In the migration from shared prefs to the official language APIs, we
did not update this screen to let the preference row know where to now
fetch this information. So instead what it did is that it looked at
shared prefs, which we've cleared in the migration, and didn't find
anything.
Then when selecting a new language, it'd set the shared prefs again,
something that we don't want, since it'd then trigger the migration
again on next launch in LanguageService::performOneTimeLanguageMigration
The two changes I'm introducing here, are a quick fix without changing
this screen too much, but letting languagePreference update its value
using the LanguageService, which uses the official language APIs under
the hood to fetch the chosen language.
And the change from `true` to `false` inside the
setOnPreferenceChangeListener is so that it does not actually try and
update the shared preferences again. But suffices to simply run
`viewModel.applyLanguage(language)` which does properly update the
LanguageService language anyway.