Remove all references to isViewingOffer
We do not want to move to the offer by persisting the quoteCartId.
Besides that, if the app was killed, and the app icon was clicked again,
it would not navigate to the offer since the order of precedence for
getLoginStatus was so that if isLoggedIn is true, it precedes checking
if one was viewing an offer. And going forward, we never want to go to
the offer page if not already logged in. Therefore the past behavior is
preserved (after we switched to web onboarding of course), while
removing a bunch of code.
In the future, preferably, the app should jump back to the offer page by
properly handling process death, and not by relying on the splash screen
to read shared prefs to make such a decision.