Make AuthTokenStorage save expiration date as an Instant
Save the ISO8601 string representation as it's a lossless conversion
going to and from it. While epochMilliseconds loses some
sub-millisecond precision which makes the tests more awkward.

Create the `Local*` alternatives of the tokens which store the Instant
instead of the seconds until expiration which are returned from the
backend which are not really useful at any point after as soon as we
receive them.