Add AuthTokenRefreshingInterceptor
Handles the entire process of checking the current tokens, refreshing
them if the access token is expired, and clearing them if the refresh
token is also expired.

Make AuthTokenService work so that all the suspending functions fetch
directly from storage, ensuring no possible race conditions, while
keeping the `authStatus` StateFlow to be used for reactive purposes
(logging out) or for a non-blocking fast read on the interceptor in the
happy path where the access token simply exists and is not expired.