mutex lock only the token acquisition part
Having the chain.proceed() inside the lock meant that it was locking
until the entire request was finished and *then* the lock was released,
making the request unnecessarily slow. The only thing we need to be
locked behind the mutex is the token acquisition, so this change does
this exactly, and continues with the token outside of the lock.