Do nothing if the value passed inside showKeyboard is null
The way it was handled before, it seems like the reference to the view
is captured, but the coroutine runs later at which point it may be null
again. Or it is never found in the first place, but we cast to non-null
but it findViewById can in fact return null.

This guards against both cases by searching for the view only when the
coroutine actually runs, but also guards against null by simply not
showing the keyboard.
Better not to show the keyboard than to crash the app in such cases.