Adopt m3 theme using our existing colors
Do some work on providing a sane mapping of what we were mapping our
colors to the m3 color scheme. Described thoroughly in the README.md of
the core-design-system module.
Along with the colorscheme, configure the typography and shapes.
In this process, also cut some unused colors from the xml file