Replace all deprecated onBackPressed
Replacement is going through the onBackPressedDispatcher. This is a
backwards compatible API to allow calling the new
onBackInvokedDispatcher on newer API levels as one should.

The way that one works, is that if there is one registered, the real
back press will not happen. Instead, what we put in the lambda will
trigger. So we are then responsible for either intercepting it and
doing what we need, or if we do not