Update letter spacings everywhere
Figma using percentage based numbers means that we were using these
slightly wrong before.
+1 in Figma, is in *EM* not *sp* which is very important, which in code
translates to 0.01f.em. For -0.25 in figma, it translates to 0.0025f.em
and so on.

Also make sure to specify the font weights in the xml theme too, at
least in API >28. Not a big deal anyway, since most of them are 400
already, but this is to achieve parity with the compose theme and how it
 is specified in Figma.