Introduce :apollo:di to wire both octopus and giraffe modules to DI
This allows for :app module to not directly depend on the apollo:octopus
module, which means that it won't be able to accidentally grab the
octopus client, nor the octopus generated GQL classes.
Now future modules can depend on the one that they need, and be able to
grab the [giraffe|octopus]Client qualifier from in there to grab the
correct ApolloClient.