Little touches to the coverage screen.
Add error screen to the coverage information
Handle insets properly for all orientations + in the error screen
Rename "MoreInfo" to "InsurableLimit" to match expectations from other
parts of the app.
Do not pass unstable List<> and prefer ImmutableList instead for
composables.
Remove firebase's internal ImmutableList to not pollute the namespace.