Make SingleItemViewModel handle picking both models and brands
Do so by splitting the item brands UI state and the item models UI state
in their own StateFlow, which will contain the source of truth of which
brands and models exist. Then the uiState is derived from those (and
what it already had before), by taking the potentially selected brand
and filtering the shown models accordingly.
SingleItemViewModelTest describes and ensures this interaction works as
expected.

Also remove the resource dependency from the ViewModel to facilitate
testing.