Fix deletion step to exit when pressing Cancel button
The mutation takes in a `confirmed` parameter, but turns out it is not
used in a meaningful way, and is instead ignored completely. It's there
to fulfill some limitations that the codegen has.
This commit adjusts our end to call the mutation always with `true`, and
instead exit the flow completely if the cancel button is pressed.