Fix app resetting when opening it from home screen
The combination of FLAG_ACTIVITY_BROUGHT_TO_FRONT and
FLAG_ACTIVITY_RESET_TASK_IF_NEEDED is true when specifically, the app is
opened by clicking the home screen icon, while the app was already
running and was present in the "recents" view.

All this due to our multi-activity setup, so if we ever manage to use a
single activity setup, or at least handle the login screen along with
the loggedIn screen in one activity, we can remove this workaround.