Replace chat TextField with compose
This solves two issues:
The text field can now expand appropriately vertically, to allow the
user to see the entire text that they are writing.
The clickable icon on the right side is now properly clickable with the
correct touch size and a ripple to indicate when it's clicked. This also
helps with the TextField text not to go too close to that icon.

As a side effect of this, make those items be 48x48 instead of 40x40, to
 make them look the same size by default, and follow proper touch size
 standards.