Remove unused FileUploadOutcome
Was only used to get a report that it's done, without using the result
in any way other than informing the UI that the operation is done. So
changed the model to reflect that fact.

Before, if it failed,
`attachPickerDialog?.uploadingTakenPicture(false)` would never be
called, and it'd be stuck on a loading state forever.