Remove a bunch of unused stuff
The entire engineering directory.
The used parts which was the impersonation feature is now moved just to
the debug directory instead.
Remove engineering hanalytics tracking
This was for Martyna to be able while running the app to see which
events have happened. I’ve never used it even once myself.
Remove all mock activities. Used it last time over a year ago and never
again
Remove all references to mixpanel which is unused for a long time now
too
And some unused styles/resources