Add more home model items
Bunch common claims in one item so that they can be rendered in a grid
view.
This is in the efforts to recreate the custom stuff that existed in the
HomeItemDecoration which is a bit harder to translate over.
Also move ChangeAddress, ClaimStatus, ConectPayin and Space.
Some are easier as they were in compose already anyway.