Fix home items for common claims
Tests caught the situations where we thought we wouldn't have common
claims when the feature flag was off, but travel certificate also uses
common claims to show their card.
This change fixes it so that we consider both things before knowing if
we want to render that section or not at all.