Remove extra fragments, inline screen in the activity
Remove the error fragment and show it when there's an error locally.
Inline the functionality of the IdentityInputFragment too. Shown when
there is no error.

This also fixes the error where if a login process had started, and they
it's cancelled, when you come back to the app, you can now start it
again without having to leave this screen and come in again.