Make gradient animate and fix inset consuming
When there is a known background gradient, first show it immediately
and from then on animate between them when it changes.
Also consume the correct insets only from the content depending on if
the nav rail or the bottom nav is showing.
And flip the order of the column/row of the navrail/bottombar so that
the bottom bar coming in with its own background feels more natural.