Move home and home related screens to :app:feature:home
In this process, a lot of XML files were also moved into the feature
module, which required moving some of the View related extensions to
common modules like core:common-android
Some drawable resources were also pulled into core-resources

All base_margin_X are replaced with just xdp, since they were not going
to be altered ever anyway. They always will be what they are specified
as.

Replace CommonClaimActivity with a simple compose CommonClaimDestination
Same with ChangeAddressActivity with a compose ChangeAddressDestination
Rename ChangeAddressActivity to LegacyChangeAddressActivity to help with
 ambiguity between the two.

To do these transformations to using more androidx.navigation, all top
level destinations also now have a navigation graph attached to them.
This allows to be able to check the hierarchy for that
TopLevelGraph in order to know in which hierarchy we're in. This is used
to highlight the correct navigation bar item.